+++ semaphore-support/src/App.jsx
@@ -8,6 +8,7 @@ import { createBrowserRouter, createRoutesFromElements, Route } from "react-rout
 import Layout from "./Layout";
 import ComingSoon from "./pages/ComingSoon";
 import { RouterProvider } from "react-router-dom";
+import HeadsInfo from "./pages/HeadsInfo";
 const App = () => {
  const router = createBrowserRouter(
     createRoutesFromElements(
@@ -16,7 +17,7 @@ const App = () => {
         <Route path="FAQ" element={<Faq />} />
         <Route path="Events" element={<Rules />} />
         <Route path="Timings" element={<ComingSoon />} />
-        <Route path="HeadsAndCoordinators" element={<ComingSoon />} />
+        <Route path="HeadsAndCoordinators" element={<HeadsInfo />} />
       </Route>
     )
   );
