+++ semaphore-support/src/pages/HeadsInfo.jsx
@@ -1,0 +1,101 @@
+import { useState, useMemo, useContext } from "react";
+import HeadCard from "../components/HeadDetail/HeadCard";
+import Heading from "../components/Heading/Heading";
+import { SemaphoreContext } from "../context/SemaphoreContext";
+
+export default function HeadsInfo() {
+  const [selectedEvent, setSelectedEvent] = useState("all");
+  const { titles, headsData } = useContext(SemaphoreContext);
+
+  const events = useMemo(() => {
+    const uniqueEvents = [...new Set(headsData.map((head) => head.event))];
+    return uniqueEvents.sort();
+  }, [headsData]);
+
+  const filteredHeads = useMemo(() => {
+    return selectedEvent === "all"
+      ? headsData
+      : headsData.filter((head) => head.event === selectedEvent);
+  }, [selectedEvent, headsData]);
+
+  const getInitials = (name) =>
+    name
+      .split(" ")
+      .map((word) => word[0])
+      .join("")
+      .toUpperCase();
+
+  const copyToClipboard = async (text) => {
+    try {
+      await navigator.clipboard.writeText(text);
+      alert(`Copied: ${text}`);
+    } catch (err) {
+      console.error("❌ Failed to copy to clipboard:", err);
+    }
+  };
+
+  return (
+    <div className="min-h-screen bg-dominant text-white">
+      <header className="sticky top-0 z-20 bg-dominant border-b border-accent shadow-md">
+        <div className="max-w-7xl mx-auto px-4 py-4 sm:py-6 flex flex-col sm:flex-row justify-between items-center gap-4">
+          <Heading
+            heading={titles.pages.headsPage.heading}
+            subheading={titles.pages.headsPage.subHeading}
+            className="!bg-transparent !mb-0"
+          />
+
+          <div className="flex items-center gap-2">
+            <label
+              htmlFor="event-select"
+              className="text-sm font-medium text-accent whitespace-nowrap"
+            >
+              Filter by:
+            </label>
+            <select
+              id="event-select"
+              value={selectedEvent}
+              onChange={(e) => setSelectedEvent(e.target.value)}
+              className="px-3 py-2 bg-dominant text-accent border border-accent rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-accent text-sm min-w-[150px] sm:min-w-[180px]"
+            >
+              <option value="all">All Events ({headsData.length})</option>
+              {events.map((event) => {
+                const count = headsData.filter((h) => h.event === event).length;
+                return (
+                  <option key={event} value={event}>
+                    {event} ({count})
+                  </option>
+                );
+              })}
+            </select>
+          </div>
+        </div>
+      </header>
+
+      <main className="p-4 sm:p-6 lg:p-8">
+        <div className="max-w-7xl mx-auto">
+          <div className="mb-6 text-center">
+            <p className="text-accent text-sm">
+              Showing{" "}
+              <span className="font-semibold">{filteredHeads.length}</span>{" "}
+              {selectedEvent === "all"
+                ? "event heads"
+                : `${selectedEvent} head${filteredHeads.length !== 1 ? "s" : ""
+                }`}
+            </p>
+          </div>
+
+          <div className="grid gap-6 sm:gap-8 grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+            {filteredHeads.map((head) => (
+              <HeadCard
+                key={head.id}
+                head={head}
+                getInitials={getInitials}
+                copyToClipboard={copyToClipboard}
+              />
+            ))}
+          </div>
+        </div>
+      </main>
+    </div>
+  );
+}
